Include common requests component. Requests menu in slices
