+++ src/app/app-routing.module.ts
@@ -45,8 +45,8 @@ import { SlicesInstancesCreateComponent } from "./service-platform/slices-instan
 
 import { SmNetworkServicesComponent } from "./service-management/sm-network-services/sm-network-services.component";
 import { SmNetworkServicesDetailComponent } from "./service-management/sm-network-services-detail/sm-network-services-detail.component";
-import { RequestsComponent } from "./service-management/requests/requests.component";
-import { RequestDetailComponent } from "./service-management/request-detail/request-detail.component";
+import { RequestsComponent } from "./shared/components/requests/requests.component";
+import { RequestDetailComponent } from "./shared/components/request-detail/request-detail.component";
 import { NetworkServiceInstancesComponent } from "./service-management/network-service-instances/network-service-instances.component";
 import { NetworkServiceInstancesDetailComponent } from "./service-management/network-service-instances-detail/network-service-instances-detail.component";
 import { FunctionRecordsDetailComponent } from "./service-management/function-records-detail/function-records-detail.component";
@@ -187,6 +187,11 @@ const routes: Routes = [
           { path: "new", component: SlicesInstancesCreateComponent },
           { path: ":id", component: SlicesInstancesDetailComponent }
         ]
+      },
+      {
+        path: "service-platform/slices/slices-requests",
+        component: RequestsComponent,
+        children: [{ path: ":id", component: RequestDetailComponent }]
       },
       // Service Management section
       {
