+++ src/app/shared/components/requests/requests.component.html
@@ -1,4 +1,4 @@
-<h1 class="sectionTitle">SERVICE MANAGEMENT</h1>
+<h1 class="sectionTitle">{{ section }}</h1>
 
 <h2 class="subsectionTitle">Requests</h2>
