+++ src/app/shared/components/requests/requests.component.scss
@@ -1,4 +1,4 @@
-@import "../../../constants.scss";
+@import "../../../../constants.scss";
 app-requests {
   position: relative;
   display: flex;
