+++ src/app/shared/services/common/common.service.ts
@@ -9,6 +9,7 @@ import { AuthService } from "../../../authentication/auth.service";
 export class CommonService {
   authHeaders: HttpHeaders;
   request_uuid: string;
+  NA = "Not available";
 
   constructor(
     private authService: AuthService,
@@ -301,6 +302,111 @@ export class CommonService {
           reject("There was an error while fetching the network service!")
         );
     });
+  }
+
+  /**
+   * Retrieves a list of Network Service requests.
+   * Either following a search pattern or not.
+   *
+   * @param search [Optional] Request attributes that must be
+   *                          matched by the returned list of
+   *                          NS requests.
+   */
+  getNSRequests(search?): any {
+    return new Promise((resolve, reject) => {
+      let headers = this.authService.getAuthHeaders();
+      let url =
+        search != undefined
+          ? this.config.baseSP + this.config.requests + search
+          : this.config.baseSP + this.config.requests;
+
+      this.http
+        .get(url, {
+          headers: headers
+        })
+        .toPromise()
+        .then(response => {
+          if (response instanceof Array) {
+            let requests = new Array();
+            response.forEach(res => {
+              let req = this.prepareNSRequest({
+                requestId: res.id,
+                name: res.name,
+                serviceName: res["service"] ? res.service.name : this.NA,
+                type: res.request_type,
+                createdAt: this.formatUTCDate(res.created_at),
+                status: res.status
+              });
+              requests.push(req);
+            });
+            resolve(requests);
+          } else {
+            reject("There was an error while fetching the requests!");
+          }
+        })
+        .catch(err => {
+          reject("There was an error while fetching the requests!");
+        });
+    });
+  }
+
+  /**
+   * Retrieves a Network Service request by UUID
+   *
+   * @param uuid UUID of the desired NS request.
+   */
+  getOneNSRequest(uuid: string): any {
+    return new Promise((resolve, reject) => {
+      let headers = this.authService.getAuthHeaders();
+
+      this.http
+        .get(this.config.baseSP + this.config.requests + "/" + uuid, {
+          headers: headers
+        })
+        .toPromise()
+        .then(response => {
+          if (response.hasOwnProperty("id")) {
+            let request = {
+              requestUUID: response["id"],
+              name: response["name"],
+              type: response["request_type"],
+              updatedAt: this.formatUTCDate(response["updated_at"]),
+              status: response["status"],
+              slaUUID: response["sla_id"],
+              serviceVendor: response["service"]["vendor"],
+              serviceName: response["service"]["name"],
+              serviceVersion: response["service"]["version"],
+              serviceUUID: response["service"]["uuid"],
+              blacklist: response["blacklist"],
+              ingresses: response["ingresses"],
+              egresses: response["egresses"]
+            };
+            this.prepareNSRequest(request);
+            resolve(request);
+          } else {
+            reject("Unable to fetch the request record!");
+          }
+        })
+        .catch(err => reject("Unable to fetch the request record!"));
+    });
+  }
+
+  prepareNSRequest(item) {
+    const parts = item.type.split("_");
+    let str: string = "";
+
+    parts.forEach(part => {
+      str = str.concat(
+        part.charAt(0).toUpperCase() + part.slice(1).toLowerCase() + " "
+      );
+    });
+
+    item.type = str;
+
+    item.status =
+      item.status.charAt(0).toUpperCase() + item.status.slice(1).toLowerCase();
+
+    return item;
   }
 
   /**
