+++ src/app/shared/shared.module.ts
@@ -17,6 +17,8 @@ import { SpinnerComponent } from "./components/spinner/spinner.component";
 import { FeatureAvailableDirective } from "./directives/feature-available.directive";
 import { AngularMaterialModule } from "../angular-material/angular-material.module";
 import { FunctionsComponent } from "./components/functions/functions.component";
+import { RequestsComponent } from "./components/requests/requests.component";
+import { RequestDetailComponent } from "./components/request-detail/request-detail.component";
 
 export function initConfiguration(configService: ConfigService): Function {
   return () => configService.init();
@@ -30,7 +32,9 @@ export function initConfiguration(configService: ConfigService): Function {
     SelectComponent,
     SpinnerComponent,
     FeatureAvailableDirective,
-    FunctionsComponent
+    FunctionsComponent,
+    RequestsComponent,
+    RequestDetailComponent
   ],
   entryComponents: [DialogComponent],
   imports: [
@@ -47,7 +51,9 @@ export function initConfiguration(configService: ConfigService): Function {
     SelectComponent,
     SpinnerComponent,
     FeatureAvailableDirective,
-    FunctionsComponent
+    FunctionsComponent,
+    RequestsComponent,
+    RequestDetailComponent
   ],
   providers: [
     ConfigService,
